# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: LingCast 数字人直播平台 API（api-admin / api-user / api-scheduler 共用一份 OpenAPI
    文档；管理端写操作与 /users、/admin/* 需登录会话，观众端与 Worker Webhook 公开）。
  title: LingCast API
  version: 1.0.0
paths:
  /admin/change-name:
    post:
      consumes:
      - application/json
      parameters:
      - description: new display name
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Change the admin display name
      tags:
      - admin
  /admin/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: currentPassword + newPassword
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Change the admin password (current password required)
      tags:
      - admin
  /admin/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: username + password
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Admin login (HttpOnly cookie session)
      tags:
      - admin
...
swagger: "2.0"